give the tilt to the MetaTile
