+++ packages/design-system/package.json
@@ -48,6 +48,7 @@
     "react-timezone-select": "^1.3.2",
     "spacetime": "7.1.2",
     "spacetime-informal": "0.6.1",
+    "vanilla-tilt": "1.7.3",
     "vite": "^3.1.8"
   },
   "devDependencies": {}
