+++ packages/web/components/Player/PlayerTile.tsx
@@ -26,7 +26,6 @@ import {
   getPlayerName,
   getPlayerURL,
 } from 'utils/playerHelpers';
-import VanillaTilt from 'vanilla-tilt';
 
 import { DAOMembershipSmall } from './Section/PlayerMemberships';
 
@@ -38,12 +37,6 @@ type Props = {
 const MAX_BIO_LENGTH = 240;
 
 export const PlayerTile: React.FC<Props> = ({ player }) => {
-  if (typeof window !== 'undefined') {
-    const element = document.querySelectorAll('.js-tilt');
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    VanillaTilt.init(element);
-  }
   const description = getPlayerDescription(player);
   const displayDescription =
     (description?.length ?? 0) > MAX_BIO_LENGTH
@@ -63,141 +56,118 @@ export const PlayerTile: React.FC<Props> = ({ player }) => {
   }, [player]);
 
   return (
-    <div
-      className="js-tilt"
-      data-tilt
-      data-tilt-scale="1.03"
-      data-tilt-max="6"
-      data-tilt-glare="true"
-      data-tilt-max-glare="0.3"
-      data-tilt-speed="800"
-      data-tilt-easing="cubic-bezier(.03,.98,.52,.99)"
-      style={{
-        display: 'flex',
-        height: '100%',
-        width: '100%',
-        borderRadius: '8px',
-      }}
-    >
-      <LinkBox height="full" width="full">
-        <MetaTile height="full" width="full">
-          <NextLink
-            as={getPlayerURL(player)}
-            href="/player/[username]"
-            passHref
+    <LinkBox height="full" width="full">
+      <MetaTile height="full" width="full">
+        <NextLink as={getPlayerURL(player)} href="/player/[username]" passHref>
+          <Link
+            display="flex"
+            flexDirection="column"
+            height="100%"
+            _hover={{
+              textUnderline: 'none',
+            }}
           >
-            <Link
-              display="flex"
-              flexDirection="column"
-              height="100%"
-              _hover={{
-                textUnderline: 'none',
-              }}
-            >
-              <MetaTileHeader>
-                <VStack pos="relative">
-                  <Flex
-                    flexDir="column"
-                    gap={1}
-                    pos="absolute"
-                    left={-8}
-                    p={3}
-                    top={-8}
-                    background="rgba(255, 255, 255, 0.1)"
-                    backdropFilter="blur(10.5px)"
-                    borderRadius="8px"
-                  >
-                    {player.rank && (
-                      <WrapItem>
-                        <MetaTag
-                          backgroundColor={player.rank.toLowerCase()}
-                          size="md"
-                          color="blackAlpha.600"
-                        >
-                          {player.rank}
-                        </MetaTag>
-                      </WrapItem>
-                    )}
-                    <Text fontSize="sm" color="blueLight">
-                      XP: {Math.floor(player.totalXP).toLocaleString()}
-                    </Text>
-                  </Flex>
-                  <PlayerProfilePicture {...{ player }} size="xl" />
-                  <Flex px={3} w="full" pos="absolute" bottom={-6} zIndex={1}>
-                    <Heading
-                      size="lg"
-                      color="white"
-                      bgColor="rgba(255, 255, 255, 0.06)"
-                      style={{ backdropFilter: 'blur(10px)' }}
-                      lineHeight={1.8}
-                      justifyContent="center"
-                      px={3}
-                      width="full"
-                      textAlign="center"
-                      borderRadius={10}
-                      fontFamily="body"
-                      fontWeight={400}
-                    >
-                      {getPlayerName(player)}
-                    </Heading>
-                  </Flex>
-                </VStack>
-              </MetaTileHeader>
-              <MetaTileBody pos="relative" height="full">
-                {displayDescription && (
-                  <VStack spacing={2} align="stretch">
-                    <Text textStyle="caption">ABOUT</Text>
-                    <Text fontSize="sm">{displayDescription}</Text>
-                  </VStack>
-                )}
-                {player.skills?.length && (
-                  <VStack spacing={2} align="stretch">
-                    <Text textStyle="caption">SKILLS</Text>
-                    <SkillsTags
-                      skills={
-                        player.skills.map(
-                          ({ Skill: skill }) => skill,
-                        ) as Skill[]
-                      }
-                    />
-                  </VStack>
-                )}
-
-                <PlayerTileMemberships {...{ player }} />
-
-                <Flex justifyContent="space-between" pointerEvents="none">
-                  {!!memberships.length && (
-                    <VStack spacing={2} align="stretch">
-                      <Text textStyle="caption">MEMBER OF</Text>
-                      <HStack mt={2} position="relative" zIndex={1}>
-                        {loading && <LoadingState mb={6} />}
-                        {!loading &&
-                          memberships
-                            .slice(0, 3)
-                            .map((membership) => (
-                              <DAOMembershipSmall
-                                {...{ membership }}
-                                key={membership.address}
-                              />
-                            ))}
-                      </HStack>
-                    </VStack>
+            <MetaTileHeader>
+              <VStack pos="relative">
+                <Flex
+                  flexDir="column"
+                  gap={1}
+                  pos="absolute"
+                  left={-8}
+                  p={3}
+                  top={-8}
+                  background="rgba(255, 255, 255, 0.1)"
+                  backdropFilter="blur(10.5px)"
+                  borderRadius="8px"
+                >
+                  {player.rank && (
+                    <WrapItem>
+                      <MetaTag
+                        backgroundColor={player.rank.toLowerCase()}
+                        size="md"
+                        color="blackAlpha.600"
+                      >
+                        {player.rank}
+                      </MetaTag>
+                    </WrapItem>
                   )}
-
-                  {!!player.accounts?.length && (
-                    <VStack spacing={2} align="stretch">
-                      <Text textStyle="caption">CONTACT</Text>
-                      <HStack mt={2} pointerEvents="all">
-                        <PlayerContacts {...{ player }} disableBrightId />
-                      </HStack>
-                    </VStack>
-                  )}
+                  <Text fontSize="sm" color="blueLight">
+                    XP: {Math.floor(player.totalXP).toLocaleString()}
+                  </Text>
                 </Flex>
-              </MetaTileBody>
-            </Link>
-          </NextLink>
-        </MetaTile>
-      </LinkBox>
-    </div>
+                <PlayerProfilePicture {...{ player }} size="xl" />
+                <Flex px={3} w="full" pos="absolute" bottom={-6} zIndex={1}>
+                  <Heading
+                    size="lg"
+                    color="white"
+                    bgColor="rgba(255, 255, 255, 0.06)"
+                    style={{ backdropFilter: 'blur(10px)' }}
+                    lineHeight={1.8}
+                    justifyContent="center"
+                    px={3}
+                    width="full"
+                    textAlign="center"
+                    borderRadius={10}
+                    fontFamily="body"
+                    fontWeight={400}
+                  >
+                    {getPlayerName(player)}
+                  </Heading>
+                </Flex>
+              </VStack>
+            </MetaTileHeader>
+            <MetaTileBody pos="relative" height="full">
+              {displayDescription && (
+                <VStack spacing={2} align="stretch">
+                  <Text textStyle="caption">ABOUT</Text>
+                  <Text fontSize="sm">{displayDescription}</Text>
+                </VStack>
+              )}
+              {player.skills?.length && (
+                <VStack spacing={2} align="stretch">
+                  <Text textStyle="caption">SKILLS</Text>
+                  <SkillsTags
+                    skills={
+                      player.skills.map(({ Skill: skill }) => skill) as Skill[]
+                    }
+                  />
+                </VStack>
+              )}
+
+              <PlayerTileMemberships {...{ player }} />
+
+              <Flex justifyContent="space-between" pointerEvents="none">
+                {!!memberships.length && (
+                  <VStack spacing={2} align="stretch">
+                    <Text textStyle="caption">MEMBER OF</Text>
+                    <HStack mt={2} position="relative" zIndex={1}>
+                      {loading && <LoadingState mb={6} />}
+                      {!loading &&
+                        memberships
+                          .slice(0, 3)
+                          .map((membership) => (
+                            <DAOMembershipSmall
+                              {...{ membership }}
+                              key={membership.address}
+                            />
+                          ))}
+                    </HStack>
+                  </VStack>
+                )}
+
+                {!!player.accounts?.length && (
+                  <VStack spacing={2} align="stretch">
+                    <Text textStyle="caption">CONTACT</Text>
+                    <HStack mt={2} pointerEvents="all">
+                      <PlayerContacts {...{ player }} disableBrightId />
+                    </HStack>
+                  </VStack>
+                )}
+              </Flex>
+            </MetaTileBody>
+          </Link>
+        </NextLink>
+      </MetaTile>
+    </LinkBox>
   );
 };
