+++ packages/web/package.json
@@ -84,7 +84,6 @@
     "sourcecred": "0.11.0",
     "swr": "1.2.2",
     "urql": "2.2.0",
-    "vanilla-tilt": "1.7.3",
     "web3.storage": "3.5.6",
     "web3modal": "1.9.5"
   },
